test: add integration test for empty volume
