+++ agent/engine/engine_sudo_linux_integ_test.go
@@ -16,16 +16,24 @@
 package engine
 
 import (
+	"io/ioutil"
+	"path/filepath"
+	"strings"
 	"testing"
 	"time"
 
 	apicontainer "github.com/aws/amazon-ecs-agent/agent/api/container"
 	apitask "github.com/aws/amazon-ecs-agent/agent/api/task"
 	"github.com/aws/amazon-ecs-agent/agent/config"
-	cgroup "github.com/aws/amazon-ecs-agent/agent/taskresource/cgroup/control"
 	"github.com/aws/amazon-ecs-agent/agent/taskresource"
+	cgroup "github.com/aws/amazon-ecs-agent/agent/taskresource/cgroup/control"
+	taskresourcevolume "github.com/aws/amazon-ecs-agent/agent/taskresource/volume"
 	"github.com/aws/amazon-ecs-agent/agent/utils/ioutilwrapper"
 	"github.com/stretchr/testify/assert"
+)
+
+const (
+	testVolumeImage = "127.0.0.1:51670/amazon/amazon-ecs-volumes-test:latest"
 )
 
 func TestStartStopWithCgroup(t *testing.T) {
@@ -77,4 +85,39 @@ func TestStartStopWithCgroup(t *testing.T) {
 	}
 	assert.False(t, ok, "Expected container to have been swept but was not")
 	assert.False(t, control.Exists(cgroupRoot))
+}
+
+func TestLocalHostVolumeMount(t *testing.T) {
+	cfg := defaultTestConfigIntegTest()
+	taskEngine, done, _ := setup(cfg, nil, t)
+	defer done()
+
+	// creates a task with local volume
+	testTask := createTestLocalVolumeMountTask("TestLocalHostVolumeMount")
+	stateChangeEvents := taskEngine.StateChangeEvents()
+	go taskEngine.AddTask(testTask)
+
+	verifyContainerRunningStateChange(t, taskEngine)
+	verifyTaskIsRunning(stateChangeEvents, testTask)
+	verifyContainerStoppedStateChange(t, taskEngine)
+	verifyTaskIsStopped(stateChangeEvents, testTask)
+
+	assert.NotNil(t, testTask.Containers[0].GetKnownExitCode(), "No exit code found")
+	assert.Equal(t, 0, *testTask.Containers[0].GetKnownExitCode(), "Wrong exit code")
+
+	assert.Contains(t, testTask.Volumes[0].Volume.SourcePath(), "/var/lib/docker")
+	data, err := ioutil.ReadFile(filepath.Join(testTask.Volumes[0].Volume.SourcePath(), "hello-from-container"))
+	assert.Nil(t, err, "Unexpected error")
+	assert.Equal(t, "empty-data-volume", strings.TrimSpace(string(data)), "Incorrect file contents")
+}
+
+func createTestLocalVolumeMountTask(tmpPath string) *apitask.Task {
+	testTask := createTestTask("testLocalHostVolumeMount")
+	testTask.Volumes = []apitask.TaskVolume{{Name: "test-tmp", Volume: &taskresourcevolume.LocalVolume{}}}
+	testTask.Containers[0].Image = testVolumeImage
+	testTask.Containers[0].MountPoints = []apicontainer.MountPoint{{ContainerPath: "/host/tmp", SourceVolume: "test-tmp"}}
+	testTask.ResourcesMapUnsafe = make(map[string][]taskresource.TaskResource)
+	testTask.Containers[0].TransitionDependenciesMap = make(map[apicontainer.ContainerStatus]apicontainer.TransitionDependencySet)
+	testTask.Containers[0].Command = []string{`echo -n "empty-data-volume" > /host/tmp/hello-from-container;`}
+	return testTask
 }
