+++ agent/taskresource/types/types_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 // Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"). You may
